// reactions/sdk/javascript/src/types/ViewItem.d.ts
// Copyright 2024 The Drasi Authors.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

/**
 * Header information for a view item
 */
export interface ViewItemHeader {
  /**
   * The sequence number of the event
   */
  sequence: number;
  state?: string;
  /**
   * The time at which the source change was recorded
   */
  timestamp: number;
}

/**
 * A view item representing a single result from a query
 */
export interface ViewItem {
  header?: ViewItemHeader;
  data?: Record<string, any>;
}